+++ src/components/common/layout/style.ts
@@ -1,0 +1,7 @@
+import { styled } from 'styled-components';
+
+export const LayoutContainer = styled.div`
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 30px;
+`;
