+++ src/components/footer/Footer.tsx
@@ -1,0 +1,15 @@
+import Layout from '../common/layout/Layout';
+import * as St from './style';
+
+const Footer = () => {
+  return (
+    <St.ContainerFooter>
+      <Layout>
+        <St.FootTit>React 6</St.FootTit>
+        <St.FootInfo></St.FootInfo>
+      </Layout>
+    </St.ContainerFooter>
+  );
+};
+
+export default Footer;
