+++ src/components/footer/style.ts
@@ -1,0 +1,8 @@
+import { styled } from 'styled-components';
+
+export const ContainerFooter = styled.footer`
+  padding: 20px 0;
+  background-color: #efefef;
+`;
+export const FootTit = styled.h4``;
+export const FootInfo = styled.div``;
