+++ src/components/header/Header.tsx
@@ -1,19 +1,17 @@
-import { styled } from 'styled-components';
-import Layout from '../common/Layout';
+import Layout from '../common/layout/Layout';
 import { Link } from 'react-router-dom';
+import * as St from './style';
 
 const Header = () => {
   return (
-    <StHeader>
+    <St.ContainerHeader>
       <Layout>
         <h1>
           <Link to={'/'}></Link>
         </h1>
       </Layout>
-    </StHeader>
+    </St.ContainerHeader>
   );
 };
 
 export default Header;
-
-const StHeader = styled.header``;
