+++ src/components/header/style.ts
@@ -1,0 +1,3 @@
+import { styled } from 'styled-components';
+
+export const ContainerHeader = styled.header``;
